LayoutTests/imported/w3c/web-platform-tests/payment-request/interfaces.https-expected.txt

CONSOLE MESSAGE: line 251: callback not yet supported

FAIL Setup for Payment Request API IDL tests. promise_test: Unhandled rejection with value: object "Unrecognised tokens, line 1 (tokens: '{"error": {')
[
    {
        "type": "other",
        "value": "{"
    },
    {
        "type": "string",
        "value": "\"error\""
    },
    {
        "type": "other",
        "value": ":"
    },
    {
        "type": "whitespace",
        "value": " "
    },
    {
        "type": "other",
        "value": "{"
    }
]"
